requirements.txt
colorutils==0.3.0
extcolors==1.0.0
imutils==0.5.4
lmfit==1.2.2
matplotlib==3.5.3
jinja2==3.1.2
numpy==1.24.3
opencv-python-headless==4.8.0.76
openpyxl==3.0.10
pandas==1.5.3
Pillow==9.4.0
plotly==5.9.0
git+https://github.com/PMEAL/porespy.git@dev
scikit-image==0.12.0
scikit-learn==1.3.0
scipy==1.10.1
simplification==0.6.11
statsmodels==0.14.0
streamlit==1.26.0
streamlit-aggrid==0.3.4.post3
streamlit-drawable-canvas==0.8.0